requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
pydantic
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
alembic
email_validator
cryptography
python-dotenv        
python-multipart     
pydantic-settings    
psycopg2-binary
